test(dedupe): consolidate ws_events tests (5 → 4 files)
Cluster c00dd1e7c53c: merge similar_names test files for ws_events

Files consolidated:
- test_ws_events.py (NEW): merged types_and_sequence_guard + health (129 lines)
- test_ws_events_dispatcher.py: kept (128 lines)
- test_ws_events_parsing.py: kept (197 lines)
- test_ws_events_edge.py: renamed from edges (142 lines)

Removed:
- test_ws_events_health.py
- test_ws_events_types_and_sequence_guard.py

Note: Only -1 file reduction (vs expected -4) due to 240-line limit.
All 33 tests pass.
