+++ tests/_triage/dedupe_candidates.yaml
@@ -2501,7 +2501,9 @@ clusters:
     expected_file_delta: -4
     priority: low
     rationale: Consolidate 5 test files into single parametrized file
-    status: pending
+    status: done
+    owner: claude
+    pr_url: https://github.com/Solders-Girdles/GPT-Trader/pull/317
     added: '2026-01-20'
   c1e4c1ed0ffb:
     type: similar_names
