+++ tests/unit/gpt_trader/features/brokerages/coinbase/test_ws_events_edge.py
